Refactor: Consolidate health check framework into health/checks/ package
- Create health/checks/ package directory
- Move DatabaseHealthCheck -> health/checks/database.py
- Move OpenAIHealthCheck -> health/checks/openai.py
- Move TwitchAPIHealthCheck -> health/checks/twitch.py
- Move locale_file_health_check -> health/checks/locales.py
- Move external_api_health_checks -> health/checks/external.py
- Move health/checks.py (base classes) -> health/checks/__init__.py
- Update health/__init__.py to re-export all check classes
- Update main.py imports to use health package
- Update health_check.py (compat shim) to re-export from health package
- Update extensions/health_check.py imports
- Remove old root-level health check files

Closes #1613
