+++ extensions/health_check.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from health_check import HealthCheck, HealthCheckResult, HealthStatus
+from health.checks import HealthCheck, HealthCheckResult, HealthStatus
 
 
 class BackgroundLoopHealthCheck(HealthCheck):
