+++ health/__init__.py
@@ -5,13 +5,33 @@ for startup validation and periodic health monitoring.
 """
 
 from health.checks import HealthCheck, HealthCheckResult, HealthStatus
+from health.checks.database import DatabaseHealthCheck
+from health.checks.external import (
+    BrawlStarsHealthCheck,
+    BytebinHealthCheck,
+    GIPHYHealthCheck,
+    GitHubAPIHealthCheck,
+    ImgBBHealthCheck,
+)
+from health.checks.locales import LocaleFileHealthCheck
+from health.checks.openai import OpenAIHealthCheck
+from health.checks.twitch import TwitchAPIHealthCheck
 from health.manager import HealthCheckManager
 from health.notifier import notify_health_failures
 
 __all__ = [
+    "BrawlStarsHealthCheck",
+    "BytebinHealthCheck",
+    "DatabaseHealthCheck",
+    "GIPHYHealthCheck",
+    "GitHubAPIHealthCheck",
     "HealthCheck",
     "HealthCheckResult",
     "HealthStatus",
     "HealthCheckManager",
+    "ImgBBHealthCheck",
+    "LocaleFileHealthCheck",
+    "OpenAIHealthCheck",
+    "TwitchAPIHealthCheck",
     "notify_health_failures",
 ]
